+++ DNN Platform/Modules/ResourceManager/ResourceManager.Web/src/components.d.ts
@@ -5,11 +5,14 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
+import { GetFolderContentResponse, Item } from "./services/ItemsClient";
 import { FolderTreeItem } from "./services/InternalServicesClient";
-import { GetFolderContentResponse } from "./services/ItemsClient";
 export namespace Components {
     interface DnnActionCreateFolder {
         "parentFolderId": number;
+    }
+    interface DnnActionEditItem {
+        "item": Item;
     }
     interface DnnResourceManager {
         /**
@@ -19,7 +22,7 @@ export namespace Components {
     }
     interface DnnRmActionsBar {
     }
-    interface DnnRmEditFolder {
+    interface DnnRmCreateFolder {
     }
     interface DnnRmFilesPane {
         /**
@@ -29,9 +32,9 @@ export namespace Components {
     }
     interface DnnRmFolderContextMenu {
         /**
-          * The ID of the folder onto which the context menu was triggered on.
+          * The item that triggered this menu.
          */
-        "clickedFolderId": number;
+        "item": Item;
     }
     interface DnnRmFolderList {
     }
@@ -44,6 +47,10 @@ export namespace Components {
           * The basic information about the folder
          */
         "folder": FolderTreeItem;
+        /**
+          * The ID of the parent folder.
+         */
+        "parentFolderId": number;
     }
     interface DnnRmItemsCardview {
         /**
@@ -73,6 +80,12 @@ declare global {
         prototype: HTMLDnnActionCreateFolderElement;
         new (): HTMLDnnActionCreateFolderElement;
     };
+    interface HTMLDnnActionEditItemElement extends Components.DnnActionEditItem, HTMLStencilElement {
+    }
+    var HTMLDnnActionEditItemElement: {
+        prototype: HTMLDnnActionEditItemElement;
+        new (): HTMLDnnActionEditItemElement;
+    };
     interface HTMLDnnResourceManagerElement extends Components.DnnResourceManager, HTMLStencilElement {
     }
     var HTMLDnnResourceManagerElement: {
@@ -85,11 +98,11 @@ declare global {
         prototype: HTMLDnnRmActionsBarElement;
         new (): HTMLDnnRmActionsBarElement;
     };
-    interface HTMLDnnRmEditFolderElement extends Components.DnnRmEditFolder, HTMLStencilElement {
+    interface HTMLDnnRmCreateFolderElement extends Components.DnnRmCreateFolder, HTMLStencilElement {
     }
-    var HTMLDnnRmEditFolderElement: {
-        prototype: HTMLDnnRmEditFolderElement;
-        new (): HTMLDnnRmEditFolderElement;
+    var HTMLDnnRmCreateFolderElement: {
+        prototype: HTMLDnnRmCreateFolderElement;
+        new (): HTMLDnnRmCreateFolderElement;
     };
     interface HTMLDnnRmFilesPaneElement extends Components.DnnRmFilesPane, HTMLStencilElement {
     }
@@ -153,9 +166,10 @@ declare global {
     };
     interface HTMLElementTagNameMap {
         "dnn-action-create-folder": HTMLDnnActionCreateFolderElement;
+        "dnn-action-edit-item": HTMLDnnActionEditItemElement;
         "dnn-resource-manager": HTMLDnnResourceManagerElement;
         "dnn-rm-actions-bar": HTMLDnnRmActionsBarElement;
-        "dnn-rm-edit-folder": HTMLDnnRmEditFolderElement;
+        "dnn-rm-create-folder": HTMLDnnRmCreateFolderElement;
         "dnn-rm-files-pane": HTMLDnnRmFilesPaneElement;
         "dnn-rm-folder-context-menu": HTMLDnnRmFolderContextMenuElement;
         "dnn-rm-folder-list": HTMLDnnRmFolderListElement;
@@ -172,6 +186,9 @@ declare namespace LocalJSX {
     interface DnnActionCreateFolder {
         "parentFolderId"?: number;
     }
+    interface DnnActionEditItem {
+        "item": Item;
+    }
     interface DnnResourceManager {
         /**
           * The ID of the module.
@@ -180,7 +197,7 @@ declare namespace LocalJSX {
     }
     interface DnnRmActionsBar {
     }
-    interface DnnRmEditFolder {
+    interface DnnRmCreateFolder {
         /**
           * Fires when there is a possibility that some folders have changed. Can be used to force parts of the UI to refresh.
          */
@@ -194,9 +211,9 @@ declare namespace LocalJSX {
     }
     interface DnnRmFolderContextMenu {
         /**
-          * The ID of the folder onto which the context menu was triggered on.
+          * The item that triggered this menu.
          */
-        "clickedFolderId": number;
+        "item": Item;
     }
     interface DnnRmFolderList {
     }
@@ -213,6 +230,10 @@ declare namespace LocalJSX {
           * Fires when a context menu is opened for this item. Emits the folder ID.
          */
         "onDnnRmcontextMenuOpened"?: (event: CustomEvent<number>) => void;
+        /**
+          * The ID of the parent folder.
+         */
+        "parentFolderId": number;
     }
     interface DnnRmItemsCardview {
         /**
@@ -236,9 +257,10 @@ declare namespace LocalJSX {
     }
     interface IntrinsicElements {
         "dnn-action-create-folder": DnnActionCreateFolder;
+        "dnn-action-edit-item": DnnActionEditItem;
         "dnn-resource-manager": DnnResourceManager;
         "dnn-rm-actions-bar": DnnRmActionsBar;
-        "dnn-rm-edit-folder": DnnRmEditFolder;
+        "dnn-rm-create-folder": DnnRmCreateFolder;
         "dnn-rm-files-pane": DnnRmFilesPane;
         "dnn-rm-folder-context-menu": DnnRmFolderContextMenu;
         "dnn-rm-folder-list": DnnRmFolderList;
@@ -256,9 +278,10 @@ declare module "@stencil/core" {
     export namespace JSX {
         interface IntrinsicElements {
             "dnn-action-create-folder": LocalJSX.DnnActionCreateFolder & JSXBase.HTMLAttributes<HTMLDnnActionCreateFolderElement>;
+            "dnn-action-edit-item": LocalJSX.DnnActionEditItem & JSXBase.HTMLAttributes<HTMLDnnActionEditItemElement>;
             "dnn-resource-manager": LocalJSX.DnnResourceManager & JSXBase.HTMLAttributes<HTMLDnnResourceManagerElement>;
             "dnn-rm-actions-bar": LocalJSX.DnnRmActionsBar & JSXBase.HTMLAttributes<HTMLDnnRmActionsBarElement>;
-            "dnn-rm-edit-folder": LocalJSX.DnnRmEditFolder & JSXBase.HTMLAttributes<HTMLDnnRmEditFolderElement>;
+            "dnn-rm-create-folder": LocalJSX.DnnRmCreateFolder & JSXBase.HTMLAttributes<HTMLDnnRmCreateFolderElement>;
             "dnn-rm-files-pane": LocalJSX.DnnRmFilesPane & JSXBase.HTMLAttributes<HTMLDnnRmFilesPaneElement>;
             "dnn-rm-folder-context-menu": LocalJSX.DnnRmFolderContextMenu & JSXBase.HTMLAttributes<HTMLDnnRmFolderContextMenuElement>;
             "dnn-rm-folder-list": LocalJSX.DnnRmFolderList & JSXBase.HTMLAttributes<HTMLDnnRmFolderListElement>;
