+++ DNN Platform/Modules/ResourceManager/ResourceManager.Web/src/components/dnn-rm-actions-bar/readme.md
@@ -15,6 +15,7 @@
 
 - dnn-vertical-overflow-menu
 - [dnn-action-create-folder](../actions/dnn-action-create-folder)
+- [dnn-action-edit-item](../actions/dnn-action-edit-item)
 - dnn-collapsible
 
 ### Graph
@@ -22,12 +23,15 @@
 graph TD;
   dnn-rm-actions-bar --> dnn-vertical-overflow-menu
   dnn-rm-actions-bar --> dnn-action-create-folder
+  dnn-rm-actions-bar --> dnn-action-edit-item
   dnn-rm-actions-bar --> dnn-collapsible
   dnn-action-create-folder --> dnn-modal
-  dnn-action-create-folder --> dnn-rm-edit-folder
-  dnn-rm-edit-folder --> dnn-button
+  dnn-action-create-folder --> dnn-rm-create-folder
+  dnn-rm-create-folder --> dnn-button
   dnn-button --> dnn-modal
   dnn-button --> dnn-button
+  dnn-action-edit-item --> dnn-modal
+  dnn-action-edit-item --> dnn-rm-create-folder
   dnn-rm-right-pane --> dnn-rm-actions-bar
   style dnn-rm-actions-bar fill:#f9f,stroke:#333,stroke-width:4px
 ```
