+++ DNN Platform/Modules/ResourceManager/ResourceManager.Web/src/components/dnn-rm-create-folder/readme.md
@@ -17,6 +17,7 @@
 ### Used by
 
  - [dnn-action-create-folder](../actions/dnn-action-create-folder)
+ - [dnn-action-edit-item](../actions/dnn-action-edit-item)
 
 ### Depends on
 
@@ -25,11 +26,12 @@
 ### Graph
 ```mermaid
 graph TD;
-  dnn-rm-edit-folder --> dnn-button
+  dnn-rm-create-folder --> dnn-button
   dnn-button --> dnn-modal
   dnn-button --> dnn-button
-  dnn-action-create-folder --> dnn-rm-edit-folder
-  style dnn-rm-edit-folder fill:#f9f,stroke:#333,stroke-width:4px
+  dnn-action-create-folder --> dnn-rm-create-folder
+  dnn-action-edit-item --> dnn-rm-create-folder
+  style dnn-rm-create-folder fill:#f9f,stroke:#333,stroke-width:4px
 ```
 
 ----------------------------------------------
