+++ DNN Platform/Modules/ResourceManager/ResourceManager.Web/src/components/dnn-rm-left-pane/readme.md
@@ -26,11 +26,14 @@ graph TD;
   dnn-rm-folder-list-item --> dnn-rm-folder-context-menu
   dnn-treeview-item --> dnn-collapsible
   dnn-rm-folder-context-menu --> dnn-action-create-folder
+  dnn-rm-folder-context-menu --> dnn-action-edit-item
   dnn-action-create-folder --> dnn-modal
-  dnn-action-create-folder --> dnn-rm-edit-folder
-  dnn-rm-edit-folder --> dnn-button
+  dnn-action-create-folder --> dnn-rm-create-folder
+  dnn-rm-create-folder --> dnn-button
   dnn-button --> dnn-modal
   dnn-button --> dnn-button
+  dnn-action-edit-item --> dnn-modal
+  dnn-action-edit-item --> dnn-rm-create-folder
   dnn-resource-manager --> dnn-rm-left-pane
   style dnn-rm-left-pane fill:#f9f,stroke:#333,stroke-width:4px
 ```
