+++ DNN Platform/Modules/ResourceManager/ResourceManager.Web/src/components/dnn-rm-right-pane/readme.md
@@ -25,17 +25,21 @@ graph TD;
   dnn-rm-right-pane --> dnn-rm-status-bar
   dnn-rm-actions-bar --> dnn-vertical-overflow-menu
   dnn-rm-actions-bar --> dnn-action-create-folder
+  dnn-rm-actions-bar --> dnn-action-edit-item
   dnn-rm-actions-bar --> dnn-collapsible
   dnn-action-create-folder --> dnn-modal
-  dnn-action-create-folder --> dnn-rm-edit-folder
-  dnn-rm-edit-folder --> dnn-button
+  dnn-action-create-folder --> dnn-rm-create-folder
+  dnn-rm-create-folder --> dnn-button
   dnn-button --> dnn-modal
   dnn-button --> dnn-button
+  dnn-action-edit-item --> dnn-modal
+  dnn-action-edit-item --> dnn-rm-create-folder
   dnn-rm-files-pane --> dnn-rm-items-listview
   dnn-rm-files-pane --> dnn-rm-items-cardview
   dnn-rm-items-listview --> dnn-collapsible
   dnn-rm-items-listview --> dnn-rm-folder-context-menu
   dnn-rm-folder-context-menu --> dnn-action-create-folder
+  dnn-rm-folder-context-menu --> dnn-action-edit-item
   dnn-rm-items-cardview --> dnn-collapsible
   dnn-rm-items-cardview --> dnn-rm-folder-context-menu
   dnn-resource-manager --> dnn-rm-right-pane
